+++ src/typescript/Config.ts
@@ -1,4 +1,5 @@
 interface Config {
   debug: boolean;
+  showBoundingBoxes: boolean;
   touch: boolean;
 }